requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
httpx>=0.24.0
pydantic>=2.0.0
jinja2>=3.1.2
python-multipart>=0.0.6
email-validator>=2.0.0
cryptography>=40.0.0
python-dotenv>=1.0.0
aio-pika>=9.0.0
requests>=2.28.0
typing-extensions>=4.5.0 